review: split clone hints, log streaming, and disk prep out of oversized files
Pure moves: cmd/vm/run.go 690->555 (+clonehints.go), cmd/vm/lifecycle.go
514->424 (+logstream.go), hypervisor/utils.go 608->460 (+disks.go).
